Refactor SHOW READ QUERY LOAD BALANCE ALGORITHM IMPLEMENTATIONS
