+++ features/readwrite-splitting/distsql/handler/src/test/java/org/apache/shardingsphere/readwritesplitting/distsql/handler/query/ReadwriteSplittingSPIImplementationsResultRowBuilderTest.java
@@ -1,0 +1,54 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.readwritesplitting.distsql.handler.query;
+
+import org.apache.shardingsphere.infra.merge.result.impl.local.LocalDataQueryResultRow;
+import org.junit.jupiter.api.Test;
+
+import java.util.Collection;
+import java.util.Iterator;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+class ReadwriteSplittingSPIImplementationsResultRowBuilderTest {
+    
+    @Test
+    void assertGetRowData() {
+        ReadwriteSplittingSPIImplementationsResultRowBuilder rowBuilder = new ReadwriteSplittingSPIImplementationsResultRowBuilder();
+        Collection<LocalDataQueryResultRow> actual = rowBuilder.generateRows();
+        assertThat(actual.size(), is(4));
+        Iterator<LocalDataQueryResultRow> iterator = actual.iterator();
+        LocalDataQueryResultRow row = iterator.next();
+        assertThat(row.getCell(1), is("DISTSQL.FIXTURE"));
+        assertThat(row.getCell(2), is(""));
+        assertThat(row.getCell(3), is(""));
+        row = iterator.next();
+        assertThat(row.getCell(1), is("ROUND_ROBIN"));
+        assertThat(row.getCell(2), is(""));
+        assertThat(row.getCell(3), is(""));
+        row = iterator.next();
+        assertThat(row.getCell(1), is("RANDOM"));
+        assertThat(row.getCell(2), is(""));
+        assertThat(row.getCell(3), is(""));
+        row = iterator.next();
+        assertThat(row.getCell(1), is("WEIGHT"));
+        assertThat(row.getCell(2), is(""));
+        assertThat(row.getCell(3), is(""));
+    }
+}
