--- infra/distsql-handler/src/main/java/org/apache/shardingsphere/distsql/handler/type/query/rql/ShowSPIImplementationsBuilder.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.distsql.handler.type.query.rql;
-
-import org.apache.shardingsphere.infra.merge.result.impl.local.LocalDataQueryResultRow;
-import org.apache.shardingsphere.infra.spi.annotation.SingletonSPI;
-import org.apache.shardingsphere.infra.spi.type.typed.TypedSPI;
-
-import java.util.Collection;
-
-/**
- * Show SPI implementations builder.
- */
-@SingletonSPI
-public interface ShowSPIImplementationsBuilder extends TypedSPI {
-    
-    /**
-     * Generate rows.
-     *
-     * @return generated rows
-     */
-    Collection<LocalDataQueryResultRow> generateRows();
-    
-    /**
-     * Get column names.
-     *
-     * @return column names
-     */
-    Collection<String> getColumnNames();
-    
-    @Override
-    String getType();
-}
